requirements.txt
# Web Framework
flask==3.1.2
flask-sqlalchemy==3.1.1
flask-restx==1.3.2

# Image Processing
Pillow==10.4.0
opencv-python==4.12.0.88

# Machine Learning
torch==2.9.1
torchvision==0.24.1
ultralytics==8.3.233

# Data Processing
numpy==2.2.6
#pandas==2.0.3

# Database
sqlalchemy==2.0.44
psycopg2-binary==2.9.11

# Data Versioning
dvc==3.64.0

# Testing
pytest==9.0.1
pytest-cov==7.0.0

# Development Tools
requests==2.32.5
pylint==4.0.3
ruff==0.14.7
